* (bug 15461) Make IE8 turn off content sniffing. Everbody else should ignore this
Adding X-Content-Type-Options: nosniff header in WebStart.php so that it's *always* set, even for alternate entry points and when $wgOut gets disabled or overridden.

Note that this of course doesn't apply to uploaded images and such loaded directly -- the usual caveats still apply.

Have not tested to confirm that this actually protects against sniffing errors -- please test on IE8 and IE9.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@87997 dd0e9695-b195-4be7-bd10-2dea1a65a6b6