* (bug 29107) Fix regression in ResourceLoader debug mode for WikiEditor
One of WikiEditor's modules had only messages, no scripts; updates to the debug mode loader had ended up failing in the case where no scripts got passed in (if passed with a loader function we were fine, hence non-debug mode being ok)
This commit explicitly checks for the empty-array case and marks the module as ready immediately, instead of waiting for the last item in the loop to finish, which never happens. :)

Also consolidated three calls to the same few lines of code into a lambda function.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@88706 dd0e9695-b195-4be7-bd10-2dea1a65a6b6