* (bug 18003) Fix API doc help for action=opensearch with OpenSearchXml extension
Now adds the full description of the parameter and accepted values instead of just sticking it in with no validation or description.
Unfortunately still can't just fetch the param as normal, as something in ApiBase is overriding the default selection and ends up 'xmlfm' instead of 'json' that way. Very annoying!

This should do for now.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@88766 dd0e9695-b195-4be7-bd10-2dea1a65a6b6