* (bug 27465) Fix metadata extraction for SVG files using unusual namespace names
The previous fix (in r82307) only checked explicitly for a namespace given the 'svg' prefix; this fix use XML namespacing support on XMLReader to check for the actual namespace URI correctly.
Fixed up a test case (for RDF extraction) and added trimming on the whitespace.
Also added another test case file that doesn't use a namespace name on the root.



git-svn-id: svn+ssh://svn.wikimedia.org/svnroot/mediawiki/trunk@88871 dd0e9695-b195-4be7-bd10-2dea1a65a6b6